requirements.txt
# Core dependencies
daytona                 # Daytona SDK for managing development environments
python-dotenv==1.0.0    # Environment variable management
requests>=2.31.0        # HTTP library

# CLI and UI enhancements
rich==13.7.0           # Rich terminal output and progress bars
prompt-toolkit>=3.0.0  # Interactive CLI prompts

# Async support
asyncio                # Built-in, but listed for clarity
aiohttp>=3.9.0        # Async HTTP client

# Testing dependencies
pytest==7.4.3         # Testing framework
pytest-asyncio==0.21.1 # Async test support
pytest-mock>=3.12.0   # Mocking utilities
pytest-cov>=4.1.0     # Code coverage

# Optional: Real-time features
websockets==12.0      # WebSocket support for streaming